refactor(indexer): extract cursor/scope helpers into indexer/scope.rs
Move word_at, word_and_range_at, lines_for, word_and_qualifier_at,
infer_lambda_param_type_at, lambda_params_at, lambda_params_at_col,
find_lambda_param_decl, enclosing_class_at and the free functions
extract_class_decl_name, is_id_char, find_enclosing_call_name,
callee_to_qualifier from indexer.rs into src/indexer/scope.rs.

Move 39 corresponding tests into src/indexer/scope_tests.rs, following
the same #[path = "..."] pattern used by apply.rs, lookup.rs, etc.

Re-export is_id_char and find_enclosing_call_name from indexer.rs so
existing callers in infer/args.rs and infer/it_this.rs are unaffected.
